First version of the documentation (basic usage)
